% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_search.R
\name{citation_search_bmc}
\alias{citation_search_bmc}
\title{Search for citations in BMC}
\usage{
citation_search_bmc(identifiers)
}
\arguments{
\item{identifiers}{a vector of identifiers to be searched for}
}
\value{
tibble of matching dataset and publication identifiers
}
\description{
This function searches for citations in BMC. It requires that an API key
be obtained from [Springer](https://dev.springernature.com/) and set using
`scythe_set_key()`. Requests are throttled at one identifier every second
so as to not overload the PLOS API.
}
\examples{

identifiers <- c("10.18739/A22274", "10.18739/A2D08X", "10.5063/F1T151VR")
result <- citation_search_bmc(identifiers)

}
