% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_search.R
\name{citation_search_scopus}
\alias{citation_search_scopus}
\title{Search for citations in Scopus}
\usage{
citation_search_scopus(identifiers)
}
\arguments{
\item{identifiers}{a vector of identifiers to be searched for}
}
\value{
tibble of matching dataset and publication identifiers
}
\description{
This function searches for citations in Scopus. Requests are throttled
9 requests/second so as to not overload the Scopus API.
}
\examples{

identifiers <- c("10.18739/A22274", "10.18739/A2D08X", "10.5063/F1T151VR")
result <- citation_search_scopus(identifiers)

}
