% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_search_xdd.R
\name{citation_search_xdd}
\alias{citation_search_xdd}
\title{Search for citations in xDD}
\usage{
citation_search_xdd(identifiers)
}
\arguments{
\item{identifiers}{a vector of identifiers to be searched for, without hypertext
transfer protocol: "https://" or "http://"}
}
\value{
tibble of publications and their identifiers that contain
}
\description{
This function searches for citations in xDD. Uses 'snippets/term' function in xDD API and 
searches through all of xDD corpus (not limited to full-text documents).
}
\examples{
\dontrun{
identifiers <- c("10.18739/A22274", "10.18739/A2D08X", "10.5063/F1T151VR", "10.18739/A29K97")
result <- citation_search_xdd(identifiers)
}

}
