% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_test_doi.R
\name{citation_test_doi}
\alias{citation_test_doi}
\title{Internal helper function for testing}
\usage{
citation_test_doi(library)
}
\arguments{
\item{library}{a character string of a single library source ("xdd", "plos", "scopus", "springer")}
}
\value{
tibble of search results produced by single test doi citation
}
\description{
Internal helper function for testing
}
