% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_test_doi.R
\name{get_test_doi}
\alias{get_test_doi}
\title{Internal helper function for testing}
\usage{
get_test_doi(library)
}
\arguments{
\item{source}{a character string of a single source source ("xdd", "plos", "scopus", "springer")}
}
\value{
tibble of single test doi citation known to occur in source library
}
\description{
Internal helper function for testing
}
