% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_test_doi.R
\name{get_test_key}
\alias{get_test_key}
\title{Internal helper function for testing}
\usage{
get_test_key(source)
}
\arguments{
\item{source}{a character string of a single source source ("xdd", "plos", "scopus", "springer")}
}
\value{
single character string of library/source. Will return NULL if require API key not set
}
\description{
Switch designed to help testing sources that require API keys and stop R-CMD-check
from failing from its inability to access API key.
}
