% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_search_plos.R
\name{ploscompact}
\alias{ploscompact}
\title{This function is from the `rplos` package, which is no longer maintained.}
\usage{
ploscompact(l)
}
\arguments{
\item{l}{a list}
}
\description{
This function is from the `rplos` package, which is no longer maintained.
}
