% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{report_est_wait}
\alias{report_est_wait}
\title{Report estimated wait for rate limited queries}
\usage{
report_est_wait(n_queries, wait_seconds)
}
\arguments{
\item{n_queries}{number of queries}

\item{wait_seconds}{wait time in seconds between queries}
}
\description{
Report estimated wait for rate limited queries
}
