\name{scythe-package}
\alias{scythe-package}
\alias{scythe}
\docType{package}
\title{
\packageTitle{scythe}
}
\description{
\packageDescription{scythe}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{scythe}
\packageIndices{scythe}
~~ An overview of how to use the package, including the most imp