% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scythe_get_key.R
\name{scythe_get_key}
\alias{scythe_get_key}
\title{Return which API keys are set}
\usage{
scythe_get_key()
}
\value{
list A named list of key value pairs.
}
\description{
This function looks for API keys for Scopus and Springer, and returns
them if they are set.
}
