% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_search_plos.R
\name{searchplos}
\alias{searchplos}
\title{A Modified Version of rplos::searchplos}
\usage{
searchplos(
  q = NULL,
  fl = "id",
  fq = NULL,
  sort = NULL,
  start = 0,
  limit = 10,
  sleep = 6,
  errors = "simple",
  proxy = NULL,
  callopts = list(),
  progress = NULL,
  ...
)
}
\arguments{
\item{q}{Search terms, eg: field:query}

\item{fl}{Fields to return}

\item{fq}{Fields to filter query on}

\item{sort}{Sort results according to field}

\item{start}{Record to start at for pagination}

\item{limit}{Number of results to return for pagination}

\item{sleep}{Seconds to wait between requests}

\item{errors}{One of simple or complete}

\item{proxy}{List of args for proxy connection}

\item{callopts}{Optional curl options}

\item{progress}{Optional logic for progress bar}

\item{...}{Addtl Solr arguments}
}
\description{
This function is adapted from the searchplos in the `rplos` package, which is no longer maintained.
}
