% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_search_plos.R
\name{strextract}
\alias{strextract}
\title{This function is from the `rplos` package, which is no longer maintained.}
\usage{
strextract(str, pattern)
}
\arguments{
\item{str}{A string}

\item{pattern}{A regex pattern}
}
\description{
This function is from the `rplos` package, which is no longer maintained.
}
