% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_search.R
\name{citation_search}
\alias{citation_search}
\title{Search for citations in text}
\usage{
citation_search(identifier)
}
\arguments{
\item{identifier}{a vector of identifiers to be searched for}
}
\value{
tibble of matching publication identifiers
}
\description{
Search for citations in text
}
