\name{scythe-package}
\alias{scythe-package}
\alias{scythe}
\docType{package}
\title{
\packageTitle{scythe}
}
\description{
\packageDescription{scythe}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{scythe}
\packageIndices{scythe}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{scythe}

Maintainer: \packageMaintainer{scythe}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
